refactor: seller apply request 파일 이동
