+++ src/main/kotlin/com/hoppingmall/mall/user/service/seller/SellerCommandService.kt
@@ -1,6 +1,6 @@
 package com.hoppingmall.mall.user.service.seller
 
-import com.hoppingmall.mall.user.dto.request.SellerApplyRequest
+import com.hoppingmall.mall.user.dto.request.seller.SellerApplyRequest
 
 interface SellerCommandService {
     fun apply(userId: Long, request: SellerApplyRequest)
