+++ src/main/kotlin/com/hoppingmall/mall/user/service/seller/SellerCommandServiceImpl.kt
@@ -3,7 +3,7 @@ package com.hoppingmall.mall.user.service.seller
 import com.hoppingmall.mall.user.domain.Seller
 import com.hoppingmall.mall.user.domain.repository.SellerRepository
 import com.hoppingmall.mall.user.domain.repository.UserRepository
-import com.hoppingmall.mall.user.dto.request.SellerApplyRequest
+import com.hoppingmall.mall.user.dto.request.seller.SellerApplyRequest
 import com.hoppingmall.mall.user.exception.seller.SellerBusinessNumberDuplicateException
 import com.hoppingmall.mall.user.exception.seller.SellerAlreadyAppliedException
 import com.hoppingmall.mall.user.exception.user.UserNotFoundException
