+++ src/test/kotlin/com/hoppingmall/mall/user/service/seller/SellerCommandServiceImplTest.kt
@@ -4,7 +4,7 @@ import com.hoppingmall.mall.user.domain.Seller
 import com.hoppingmall.mall.user.domain.User
 import com.hoppingmall.mall.user.domain.repository.SellerRepository
 import com.hoppingmall.mall.user.domain.repository.UserRepository
-import com.hoppingmall.mall.user.dto.request.SellerApplyRequest
+import com.hoppingmall.mall.user.dto.request.seller.SellerApplyRequest
 import com.hoppingmall.mall.user.exception.seller.SellerBusinessNumberDuplicateException
 import com.hoppingmall.mall.user.exception.seller.SellerAlreadyAppliedException
 import com.hoppingmall.mall.user.exception.user.UserNotFoundException
